refactor(daemon): split watcher runtime modules
