+++ crates/board-daemon/src/watchers/tests/sockets.rs
@@ -3,92 +3,20 @@ use std::path::{Path, PathBuf};
 use std::sync::{Arc, Mutex};
 use std::time::{Duration, Instant};
 
-use super::{
-    apply_candidate, apply_signal, check_at, classify_candidates, handle_event, AgentSignal,
-    HerdrSocketSupervisor, WatchConnector, WatchEventStream, WatchSnapshot, WatchTiming,
+use super::super::herdr::{
+    handle_event_from_socket, HerdrSocketSupervisor, WatchConnector, WatchEventStream,
+    WatchSnapshot, WatchTiming,
 };
-use crate::dispatch::{finalize_run, finalize_run_timeout};
+use super::active_daemon;
 use crate::settings::DaemonSettings;
-use crate::spawner::LocalSpawner;
-use crate::spawner::RuntimeHandle;
+use crate::spawner::{LocalSpawner, RuntimeHandle};
 use crate::state::{ActiveRun, Daemon};
 use crate::store::Store;
 use board_core::config::Config;
 use board_core::db::{Db, EnqueueRun};
-use board_core::protocol::{
-    AwaitingReason, BoardChangedReason, CardCreateParams, CardStatus, Event, RunOutcome,
-};
+use board_core::protocol::{CardCreateParams, CardStatus, RunOutcome};
 use board_herdr::{AgentStatus, HerdrError, HerdrEvent};
 use tokio::sync::{broadcast, mpsc, watch};
-
-fn active_daemon() -> (Arc<Daemon>, i64, i64, broadcast::Receiver<Event>) {
-    let config = Config {
-        idle_grace_seconds: 5,
-        ..Default::default()
-    };
-    let db = Db::open_in_memory().unwrap();
-    let card = db
-        .create_card(&CardCreateParams {
-            title: "watch".into(),
-            ..Default::default()
-        })
-        .unwrap();
-    let run = db
-        .enqueue_run_uow(&EnqueueRun {
-            card_id: card.id,
-            column_id: card.column_id,
-            harness: "pi",
-            argv_json: "[\"pi\"]",
-            prompt_snapshot: "prompt",
-            system_prompt_snapshot: None,
-            launch_spec_json: None,
-            session_id: Some("session"),
-            session: None,
-        })
-        .unwrap();
-    db.promote_run_uow(run.id, Some("w1"), Some("p1"), None)
-        .unwrap();
-
-    let (events_tx, events_rx) = broadcast::channel(16);
-    let (dispatch_tx, _dispatch_rx) = mpsc::unbounded_channel();
-    let (shutdown_tx, _shutdown_rx) = watch::channel(false);
-    let d = Arc::new(Daemon::new(
-        Store::new(db),
-        config,
-        DaemonSettings::default(),
-        PathBuf::from("/tmp/board-watch.db"),
-        PathBuf::from("/tmp/board-watch.sock"),
-        Arc::new(LocalSpawner::new()),
-        None,
-        None,
-        events_tx,
-        dispatch_tx,
-        shutdown_tx,
-    ));
-    d.sched.lock().unwrap().active.insert(
-        run.id,
-        ActiveRun {
-            card_id: card.id,
-            handle: RuntimeHandle::default(),
-            started: Instant::now(),
-            timeout_deadline: None,
-            idle_since: None,
-            awaiting_since: None,
-            is_local: false,
-            pane_id: Some("p1".into()),
-        },
-    );
-    (d, run.id, card.id, events_rx)
-}
-
-fn status(status: AgentStatus) -> HerdrEvent {
-    HerdrEvent::AgentStatusChanged {
-        pane_id: "p1".into(),
-        workspace_id: Some("w1".into()),
-        status,
-        agent: Some("pi".into()),
-    }
-}
 
 /// Build two active herdr runs sharing a pane id but living on separate
 /// session sockets. The current pane-only matcher is deliberately used to
@@ -256,7 +184,7 @@ fn event_polled_from_socket(d: &Arc<Daemon>, source_socket: &PathBuf, event: Her
         .unwrap()
         .panes_by_socket
         .contains_key(source_socket));
-    super::handle_event_from_socket(d, source_socket, event);
+    handle_event_from_socket(d, source_socket, event);
 }
 
 fn shared_status(status: AgentStatus) -> HerdrEvent {
@@ -613,372 +541,5 @@ fn socket_a_pane_exit_does_not_finalize_socket_b_duplicate_pane() {
         db.get_card(other_card).unwrap().unwrap().status,
         CardStatus::Running,
         "socket A's pane exit must not finalize socket B's run",
-    );
-}
-
-#[test]
-fn working_restores_running_and_clears_idle_state() {
-    let (d, run_id, card_id, mut events) = active_daemon();
-    d.store
-        .lock()
-        .set_card_status(card_id, CardStatus::Blocked)
-        .unwrap();
-    d.sched
-        .lock()
-        .unwrap()
-        .active
-        .get_mut(&run_id)
-        .unwrap()
-        .idle_since = Some(Instant::now());
-
-    handle_event(&d, status(AgentStatus::Working));
-
-    assert_eq!(
-        d.store.lock().get_card(card_id).unwrap().unwrap().status,
-        CardStatus::Running
-    );
-    assert!(d
-        .sched
-        .lock()
-        .unwrap()
-        .active
-        .get(&run_id)
-        .unwrap()
-        .idle_since
-        .is_none());
-    assert!(matches!(
-        events.try_recv().unwrap(),
-        Event::BoardChanged {
-            reason: BoardChangedReason::CardUpdated,
-            card_id: Some(id),
-            ..
-        } if id == card_id
-    ));
-}
-
-#[test]
-fn blocked_marks_card_and_emits_change() {
-    let (d, _run_id, card_id, mut events) = active_daemon();
-    handle_event(&d, status(AgentStatus::Blocked));
-    assert_eq!(
-        d.store.lock().get_card(card_id).unwrap().unwrap().status,
-        CardStatus::Blocked
-    );
-    assert!(matches!(
-        events.try_recv().unwrap(),
-        Event::BoardChanged {
-            reason: BoardChangedReason::RunBlocked,
-            card_id: Some(id),
-            ..
-        } if id == card_id
-    ));
-}
-
-#[test]
-fn idle_arms_grace_then_becomes_awaiting_without_sleeping() {
-    let (d, run_id, card_id, _events) = active_daemon();
-    handle_event(&d, status(AgentStatus::Idle));
-    let idle_since = d
-        .sched
-        .lock()
-        .unwrap()
-        .active
-        .get(&run_id)
-        .unwrap()
-        .idle_since
-        .unwrap();
-
-    check_at(&d, idle_since + Duration::from_secs(4));
-    assert!(d.store.lock().get_run(run_id).unwrap().ended_at.is_none());
-    check_at(&d, idle_since + Duration::from_secs(5));
-
-    // Idle past grace → awaiting, NOT lost: the run stays OPEN and the
-    // card is never auto-failed.
-    let db = d.store.lock();
-    let run = db.get_run(run_id).unwrap();
-    assert!(run.ended_at.is_none());
-    assert_eq!(run.outcome, None);
-    let card = db.get_card(card_id).unwrap().unwrap();
-    assert_eq!(card.status, CardStatus::Awaiting);
-    assert_eq!(card.awaiting_reason, Some(AwaitingReason::IdleExpired));
-}
-
-#[test]
-fn stale_idle_expiry_after_working_is_ignored_without_sleeping() {
-    let (d, run_id, card_id, mut events) = active_daemon();
-    handle_event(&d, status(AgentStatus::Idle));
-    let idle_since = d
-        .sched
-        .lock()
-        .unwrap()
-        .active
-        .get(&run_id)
-        .unwrap()
-        .idle_since
-        .unwrap();
-    let now = idle_since + Duration::from_secs(5);
-    let candidate = classify_candidates(&d, now).pop().unwrap();
-
-    // Working wins after the ticker classified the old idle period but
-    // before that candidate is applied.
-    handle_event(&d, status(AgentStatus::Working));
-    while events.try_recv().is_ok() {}
-    apply_candidate(&d, candidate, now);
-
-    assert_eq!(
-        d.store.lock().get_card(card_id).unwrap().unwrap().status,
-        CardStatus::Running
-    );
-    assert!(d
-        .sched
-        .lock()
-        .unwrap()
-        .active
-        .get(&run_id)
-        .unwrap()
-        .idle_since
-        .is_none());
-    assert!(events.try_recv().is_err());
-}
-
-#[test]
-fn herdr_done_enters_awaiting_immediately_without_grace() {
-    let (d, run_id, card_id, mut events) = active_daemon();
-    handle_event(&d, status(AgentStatus::Done));
-
-    let db = d.store.lock();
-    let card = db.get_card(card_id).unwrap().unwrap();
-    assert_eq!(card.status, CardStatus::Awaiting);
-    assert_eq!(card.awaiting_reason, Some(AwaitingReason::AgentDone));
-    assert!(db.get_run(run_id).unwrap().ended_at.is_none());
-    drop(db);
-    assert!(matches!(
-        events.try_recv().unwrap(),
-        Event::BoardChanged {
-            reason: BoardChangedReason::CardUpdated,
-            card_id: Some(id),
-            ..
-        } if id == card_id
-    ));
-    // Idle bookkeeping is disarmed while awaiting.
-    let s = d.sched.lock().unwrap();
-    let a = s.active.get(&run_id).unwrap();
-    assert!(a.idle_since.is_none());
-    assert!(a.awaiting_since.is_some());
-}
-
-#[test]
-fn working_resumes_running_from_awaiting_and_shifts_the_timeout() {
-    let (d, run_id, card_id, _events) = active_daemon();
-    let deadline = Instant::now() + Duration::from_secs(60);
-    d.sched
-        .lock()
-        .unwrap()
-        .active
-        .get_mut(&run_id)
-        .unwrap()
-        .timeout_deadline = Some(deadline);
-
-    handle_event(&d, status(AgentStatus::Done));
-    assert_eq!(
-        d.store.lock().get_card(card_id).unwrap().unwrap().status,
-        CardStatus::Awaiting
-    );
-    // Simulate review time passing while awaiting.
-    let paused = d
-        .sched
-        .lock()
-        .unwrap()
-        .active
-        .get(&run_id)
-        .unwrap()
-        .awaiting_since
-        .unwrap();
-    d.sched
-        .lock()
-        .unwrap()
-        .active
-        .get_mut(&run_id)
-        .unwrap()
-        .awaiting_since = Some(paused - Duration::from_secs(30));
-
-    handle_event(&d, status(AgentStatus::Working));
-
-    let card = d.store.lock().get_card(card_id).unwrap().unwrap();
-    assert_eq!(card.status, CardStatus::Running);
-    assert_eq!(card.awaiting_reason, None);
-    let a = d.sched.lock().unwrap().active.remove(&run_id).unwrap();
-    assert!(a.awaiting_since.is_none());
-    // The column timeout was paused: the deadline absorbed the review span.
-    assert!(a.timeout_deadline.unwrap() >= deadline + Duration::from_secs(29));
-}
-
-#[test]
-fn ticker_skips_awaiting_runs_for_both_idle_and_timeout() {
-    let (d, run_id, card_id, _events) = active_daemon();
-    handle_event(&d, status(AgentStatus::Done));
-    {
-        let mut s = d.sched.lock().unwrap();
-        let a = s.active.get_mut(&run_id).unwrap();
-        a.idle_since = Some(Instant::now() - Duration::from_secs(3600));
-        a.timeout_deadline = Some(Instant::now() - Duration::from_secs(60));
-    }
-
-    check_at(&d, Instant::now());
-
-    let db = d.store.lock();
-    let run = db.get_run(run_id).unwrap();
-    assert!(run.ended_at.is_none());
-    assert_eq!(run.outcome, None);
-    let card = db.get_card(card_id).unwrap().unwrap();
-    assert_eq!(card.status, CardStatus::Awaiting);
-    assert_eq!(card.awaiting_reason, Some(AwaitingReason::AgentDone));
-}
-
-#[test]
-fn stale_signal_after_terminal_completion_is_ignored() {
-    let (d, run_id, card_id, mut events) = active_daemon();
-    finalize_run(&d, run_id, RunOutcome::Ok, None, None, false, true).unwrap();
-    while events.try_recv().is_ok() {}
-
-    apply_signal(&d, run_id, card_id, AgentSignal::Done);
-
-    let db = d.store.lock();
-    assert_eq!(db.get_run(run_id).unwrap().outcome, Some(RunOutcome::Ok));
-    assert_eq!(
-        db.get_card(card_id).unwrap().unwrap().status,
-        CardStatus::Done
-    );
-    assert!(events.try_recv().is_err());
-}
-
-#[test]
-fn preclassified_timeout_is_rejected_after_done_enters_awaiting() {
-    let (d, run_id, card_id, _events) = active_daemon();
-    let now = Instant::now();
-    d.sched
-        .lock()
-        .unwrap()
-        .active
-        .get_mut(&run_id)
-        .unwrap()
-        .timeout_deadline = Some(now - Duration::from_secs(1));
-    assert!(d
-        .sched
-        .lock()
-        .unwrap()
-        .active
-        .get(&run_id)
-        .unwrap()
-        .timeout_deadline
-        .is_some_and(|deadline| now >= deadline));
-
-    // This signal wins after timeout classification but before its claim.
-    apply_signal(&d, run_id, card_id, AgentSignal::Done);
-    let finalized = finalize_run_timeout(
-        &d,
-        run_id,
-        now,
-        RunOutcome::Fail,
-        Some("stale timeout".into()),
-        Some("stale timeout".into()),
-        true,
-        true,
-    )
-    .unwrap();
-    assert!(finalized.is_none());
-
-    let db = d.store.lock();
-    assert!(db.get_run(run_id).unwrap().ended_at.is_none());
-    assert_eq!(
-        db.get_card(card_id).unwrap().unwrap().status,
-        CardStatus::Awaiting
-    );
-}
-
-#[test]
-fn timeout_still_finalizes_fail_when_not_awaiting() {
-    let (d, run_id, card_id, _events) = active_daemon();
-    let started = Instant::now() - Duration::from_secs(120);
-    {
-        let mut s = d.sched.lock().unwrap();
-        let a = s.active.get_mut(&run_id).unwrap();
-        a.started = started;
-        a.timeout_deadline = Some(started + Duration::from_secs(60));
-    }
-
-    check_at(&d, Instant::now());
-
-    let db = d.store.lock();
-    assert_eq!(db.get_run(run_id).unwrap().outcome, Some(RunOutcome::Fail));
-    assert_eq!(
-        db.get_card(card_id).unwrap().unwrap().status,
-        CardStatus::Failed
-    );
-}
-
-#[test]
-fn stale_status_events_without_an_active_run_are_ignored() {
-    let (d, _run_id, card_id, mut events) = active_daemon();
-    handle_event(
-        &d,
-        HerdrEvent::AgentStatusChanged {
-            pane_id: "ghost-pane".into(),
-            workspace_id: Some("w1".into()),
-            status: AgentStatus::Done,
-            agent: Some("pi".into()),
-        },
-    );
-    assert_eq!(
-        d.store.lock().get_card(card_id).unwrap().unwrap().status,
-        CardStatus::Running
-    );
-    assert!(events.try_recv().is_err());
-}
-
-#[test]
-fn pane_exit_becomes_fail_without_transition() {
-    let (d, run_id, card_id, _events) = active_daemon();
-    let original_column = d.store.lock().get_card(card_id).unwrap().unwrap().column_id;
-    handle_event(
-        &d,
-        HerdrEvent::PaneExited {
-            pane_id: "p1".into(),
-            workspace_id: Some("w1".into()),
-        },
-    );
-    let db = d.store.lock();
-    assert_eq!(db.get_run(run_id).unwrap().outcome, Some(RunOutcome::Fail));
-    let card = db.get_card(card_id).unwrap().unwrap();
-    assert_eq!(card.status, CardStatus::Failed);
-    assert_eq!(card.column_id, original_column);
-}
-
-#[test]
-fn protocol17_idle_after_done_does_not_rearm_an_awaiting_run() {
-    let (d, run_id, card_id, _events) = active_daemon();
-
-    // Protocol 17 may emit the terminal turn's `done` followed by `idle`.
-    // Done is authoritative for board review; the trailing idle must not
-    // arm a second grace period while this same run remains awaiting.
-    handle_event(&d, status(AgentStatus::Done));
-    handle_event(&d, status(AgentStatus::Idle));
-
-    let db = d.store.lock();
-    let card = db.get_card(card_id).unwrap().unwrap();
-    assert_eq!(card.status, CardStatus::Awaiting);
-    assert_eq!(card.awaiting_reason, Some(AwaitingReason::AgentDone));
-    assert!(db.get_run(run_id).unwrap().ended_at.is_none());
-    drop(db);
-    assert!(
-        d.sched
-            .lock()
-            .unwrap()
-            .active
-            .get(&run_id)
-            .unwrap()
-            .idle_since
-            .is_none(),
-        "a trailing protocol-17 idle event must not rearm idle expiry while awaiting",
     );
 }
